+++ probe/src/cgo/kindling.cpp
@@ -959,6 +959,83 @@ void attach_pid(char* pid, bool is_new_start, bool is_attach, bool is_all_attach
   }
 }
 
+void attach_agent(int64_t pid, char* error_message, bool is_attach) {
+  char result_buf[1024], command[1024];
+  string attach_command_prefix;
+  if (is_attach) {
+    attach_command_prefix = "./async-profiler/jattach.sh start ";
+  } else {
+    attach_command_prefix = "./async-profiler/jattach.sh stop ";
+  }
+  attach_command_prefix.append(std::to_string(pid));
+  strcpy(command, attach_command_prefix.c_str());
+
+  FILE* fp;
+  fp = popen(command, "r");
+  if (NULL == fp) {
+    perror("popen execute failed!\n");
+    strcpy(error_message, "popen execute failed");
+    return;
+  }
+  if (is_attach) {
+    cout << "------"
+         << " start attach agent for pid " << pid << "------" << endl;
+  } else {
+    cout << "------"
+         << " start detach agent for pid " << pid << "------" << endl;
+  }
+
+  char* error_msg;
+  while (fgets(result_buf, sizeof(result_buf), fp) != NULL) {
+    if ('\n' == result_buf[strlen(result_buf) - 1]) {
+      result_buf[strlen(result_buf) - 1] = '\0';
+    }
+    error_msg = strstr(result_buf, "[ERROR] ");
+    if (error_msg) {
+      strcpy(error_message, error_msg);
+    }
+    printf("%s\r\n", result_buf);
+  }
+
+  int rc = pclose(fp);
+  if (-1 == rc) {
+    perror("close command fp failed!\n");
+    strcpy(error_message, "close command fp failed");
+    return;
+  } else {
+    printf("command:【%s】command process status:【%d】command return value:【%d】\r\n", command,
+           rc, WEXITSTATUS(rc));
+  }
+
+  if (is_attach) {
+    cout << "------end attach agent for pid " << pid << "------" << endl;
+  } else {
+    cout << "------end detach agent for pid " << pid << "------" << endl;
+  }
+}
+
+char* start_attach_agent(int64_t pid) {
+  char* error_message = (char*) malloc(1024 * sizeof(char));
+  error_message[0] = '\0';
+  if (!inspector) {
+    strcpy(error_message, "Please start profile first");
+  } else {
+    attach_agent(pid, error_message, true);
+  }
+  return error_message;
+}
+
+char* stop_attach_agent(int64_t pid) {
+  char* error_message = (char*) malloc(1024 * sizeof(char));
+  error_message[0] = '\0';
+  if (!inspector) {
+    strcpy(error_message, "Please start profile first");
+  } else {
+    attach_agent(pid, error_message, false);
+  }
+  return error_message;
+}
+
 int start_profile() {
   if (!inspector) {
     return -1;